Initial SDL/Linux port commit
There are lots of things broken in this, but it does for the most part
work.

Calls to Win32, DirectDraw, and DirectInput are replaced with equivalent
SDL 2.0 calls. There isn't really any new abstraction layer added, and
the error handling is pretty sparse.

Some notes/todos:
- Joystick input isn't supported yet.
- There's a few keyboard input bugs, notably the "No Move" key is stuck
down, so it has been disabled.
- FMOD 3.7 is ancient, hard to find, and pretty broken on my setup. It
does play something with the ALSA backend forced, but with ~30s latency.
- The case of several filenames are wrong, so data doesn't always load
properly.
- Controls aren't loaded/saved.
- Lots of itoa() calls were replaced with the standardised sprintf()
- SDL doesn't support h-flipping Surfaces (but does support flipping
textures, but doesn't support paletted textures), so there's a custom
blitting routine for things that need mirroring. It may be buggy.
- In order to do palette conversion, the game no-longer does page
flipping, but instead creates a converted RGB surface each frame, which
is uploaded to a texture, which is scaled and blitted to the screen. So
all the VBlank timing/page flipping stuff is gone/replaced. Seems to
work so far, but could be buggy.