WIP port of audio to SDL_Mixer
The version of FMOD used was old, difficult to find, and pretty broken
on my machine. SDL_Mixer is open source, and has tracker module support,
so we sub that in in place. This also means we can have 64-bit builds,
as those old fmod versions were 32-bit only.

It's optional, so cmake -DWITH_FMOD=1 should be able to build the old
version. Basically, fake_fmod.h #defines everything necessary to get the
basics working, though there are certinaly some bugs and unimplemented
things (mostly around looping). The init code is in an #ifdef in
ld3main.cpp

Note that there aren't any sound effects in the repo. I copied the sfx/
directory from the old ld3beta2.zip file, and most of the sound effects
play. The machine gun is the obvious exception.