Add ID3v2 tags to all MP3 files
Due to what is probably a bug in SDL_Mixer, it is very particular in how
it detects MP3 files, only accepting MPEG 1 formatted files. Some of the
LD3 mp3 files are MPEG 2, and so aren't detected as MP3s by SDL_Mixer.

However, SDL_Mixer also looks for ID3 tags as a magic number in the
header. These didn't exist in the LD3 sfx files, but are easy to add.
By adding an ID3 v2.3.0 tag to all the MP3s, the first three bytes of
those files are all 'ID3', and now load properly.

(This bug was most noticable in the shotgun effect not being audible.)